requirements.mobile.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.3
asgiref==3.8.1
attrs==23.2.0
binutils==0.1.dev0
black==24.4.0
certifi==2022.9.24
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.7
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
Django==5.0
django-allauth @ git+https://github.com/pennersr/django-allauth.git@05aa77c805f3332caee30d2c4e42f3b21edd2f65
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-datetime-validators==0.1
django-dirtyfields==1.9.2
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-ipware==6.0.5
django-model-utils==4.5.1
django-notifications-hq @ git+https://github.com/django-notifications/django-notifications@ddf6b59978a1fb3fb6d01a6ba0de59a85bf2459e
django-phonenumber-field==7.3.0
django-rest-knox @ git+https://github.com/jazzband/django-rest-knox.git@09403255a069c2416a5dfb889f80342a2d7b2646
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-twilio==0.14.3.3
djangorestframework==3.15.1
djangorestframework-stubs==3.14.5
drf-generic-views==0.2
drf-pagination-meta-wrap==0.2
drf-standardized-errors==0.13.0
drf-standardized-response==1.1.1
factory-boy==3.3.0
Faker==24.4.0
filelock==3.13.3
flake8==7.0.0
frozenlist==1.4.1
glibc==0.6.1
identify==2.5.36
idna==3.4
inflection==0.5.1
jsonfield==3.1.0
mccabe==0.7.0
multidict==6.0.5
mypy==1.7.1
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.34
phonenumberslite==8.13.34
pillow==10.3.0
platformdirs==4.2.0
pre-commit==3.7.0
psycopg==3.1.18
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2022.12
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==2.0.3
python-pip==1.1.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.28.1
requests-oauthlib==2.0.0
semantic-version==2.10.0
setuptools-rust==1.9.0
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
toposort==1.10
twilio==9.0.4
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.25.1
yarl==1.9.4
